+++ zcode/__init__.py
@@ -2,17 +2,17 @@
 
 """
 
-from . import plot
-from . import math
-from . import inout
+# from . import plot
+# from . import math
+# from . import inout
 from . import constants
 
 __author__ = "Luke Zoltan Kelley"
-__version__ = "0.0.9"
+__version__ = "0.0.10"
 __email__ = "lkelley@cfa.harvard.edu"
 __status__ = "Development"
 
 __all__ = ['__version__', '__author__', '__email__', '__status__']
 
-from numpy.testing import Tester
-test = Tester().test
+# from numpy.testing import Tester
+# test = Tester().test