+++ zcode/inout/inout_core.py
@@ -8,17 +8,17 @@ Classes
 Functions
 ---------
 -   bytesString              - Return a humanized string representation of a number of bytes.
--   getFileSize              - Return a human-readable size of a file or set of files.
+-   get_file_size            - Return a human-readable size of a file or set of files.
 -   countLines               - Count the number of lines in the given file.
 -   estimateLines            - Estimate the number of lines in the given file.
--   checkPath                - Create the given filepath if it doesn't already exist.
+-   check_path               - Create the given filepath if it doesn't already exist.
 -   dictToNPZ                - Save a dictionary to the given NPZ filename.
 -   npzToDict                - Convert an NPZ file to a dictionary with the same keys and values.
 -   getProgressBar           - Wrapper to create a progressbar object with default settings.
 -   combineFiles             - Concatenate the contents of input files into a single output file.
 -   checkURL                 - Check that the given url exists.
 -   promptYesNo              - Prompt the user (via CLI) for yes or no.
--   modifyFilename           - Modify the given filename.
+-   modify_filename          - Modify the given filename.
 -   mpiError                 - Raise an error through MPI and exit all processes.
 -   ascii_table              - Print a table with the given contents to output.
 -   modify_exists            - Modify the given filename if it already exists.
@@ -37,18 +37,19 @@ from datetime import datetime
 import os
 import sys
 import re
-import logging
 import warnings
 import numpy as np
 import collections
 
 from zcode import utils
 
-__all__ = ['Keys', 'MPI_TAGS', 'StreamCapture', 'bytesString', 'getFileSize', 'get_file_size',
-           'countLines', 'estimateLines',
-           'checkPath', 'dictToNPZ', 'npzToDict', 'combineFiles', 'checkURL',
-           'promptYesNo', 'modifyFilename', 'mpiError', 'ascii_table', 'modify_exists',
-           'iterable_notstring', 'str_format_dict', 'top_dir', 'underline', 'warn_with_traceback']
+__all__ = ['Keys', 'MPI_TAGS', 'StreamCapture', 'bytesString', 'get_file_size',
+           'countLines', 'estimateLines', 'modify_filename',
+           'check_path', 'dictToNPZ', 'npzToDict', 'combineFiles', 'checkURL',
+           'promptYesNo', 'mpiError', 'ascii_table', 'modify_exists',
+           'iterable_notstring', 'str_format_dict', 'top_dir', 'underline', 'warn_with_traceback',
+           # Deprecated!
+           'modifyFilename', 'checkPath']
 
 
 class _Keys_Meta(type):
@@ -194,11 +195,6 @@ def bytesString(bytes, precision=1):
     return strSize
 
 
-def getFileSize(*args, **kwargs):
-    utils.dep_warn("getFileSize", newname="get_file_size")
-    return get_file_size(*args, **kwargs)
-
-
 def get_file_size(fnames, precision=1):
     """Return a human-readable size of a file or set of files.
 
@@ -230,17 +226,22 @@ def countLines(files, progress=False):
     """
 
     # If string, or otherwise not-iterable, convert to list
-    if(np.iterable(files) and not isinstance(files, str)): files = [files]
+    if np.iterable(files) and not isinstance(files, str):
+        files = [files]
 
-    if(progress): pbar = getProgressBar(len(files))
+    if progress:
+        warnings.warn("`progress` argument is deprecated!")
+        # pbar = getProgressBar(len(files))
 
     nums = 0
     # Iterate over each file, count lines
     for ii, fil in enumerate(files):
         nums += sum(1 for line in open(fil))
-        if(progress): pbar.update(ii)
+        # if(progress):
+        #     pbar.update(ii)
 
-    if(progress): pbar.finish()
+    # if(progress):
+    #     pbar.finish()
 
     return nums
 
@@ -273,7 +274,12 @@ def estimateLines(files):
     return numLines
 
 
-def checkPath(tpath, create=True):
+def checkPath(*args, **kwargs):
+    utils.dep_warn("checkPath", newname="check_path")
+    return check_path(*args, **kwargs)
+
+
+def check_path(tpath, create=True):
     """Create the given filepath if it doesn't already exist.
 
     Arguments
@@ -313,7 +319,7 @@ def dictToNPZ(dataDict, savefile, verbose=False, log=None):
     If ``verbose`` is True, the saved file size is printed out.
     """
     # Make sure path to file exists
-    checkPath(savefile)
+    check_path(savefile)
     # Make sure there are no scalars in the input dictionary
     for key, item in dataDict.items():
         if np.isscalar(item):
@@ -326,8 +332,8 @@ def dictToNPZ(dataDict, savefile, verbose=False, log=None):
     if not os.path.exists(savefile):
         raise RuntimeError("Could not save to file '%s'." % (savefile))
 
-    logStr = " - Saved dictionary to '%s'" % (savefile)
-    logStr += " - - Size '%s'" % (getFileSize(savefile))
+    logStr = " - Saved dictionary to '{}'".format(savefile)
+    logStr += " - - Size '{}'".format(get_file_size(savefile))
     try:
         log.debug(logStr)
     except Exception:
@@ -402,31 +408,36 @@ def combineFiles(inFilenames, outFilename, verbose=False):
     """
 
     # Make sure outfile path exists
-    checkPath(outFilename)
+    check_path(outFilename)
     inSize = 0.0
-    nums = len(inFilenames)
+    # nums = len(inFilenames)
 
     # Open output file for writing
-    if verbose: pbar = getProgressBar(nums)
+    if verbose:
+        warnings.warn("`progress` is deprecated!")
+        # pbar = getProgressBar(nums)
     with open(outFilename, 'w') as outfil:
 
         # Iterate over input files
         for ii, inname in enumerate(inFilenames):
             inSize += os.path.getsize(inname)
-            if verbose: pbar.update(ii)
+            # if verbose:
+            #     pbar.update(ii)
 
             # Open input file for reading
             with open(inname, 'r') as infil:
                 # Iterate over input file lines
                 for line in infil: outfil.write(line)
 
-    if verbose: pbar.finish()
+    # if verbose:
+    #     pbar.finish()
 
     outSize = os.path.getsize(outFilename)
     inStr = bytesString(inSize)
     outStr = bytesString(outSize)
 
-    if verbose: print("Total input size = %s, output size = %s" % (inStr, outStr))
+    if verbose:
+        print("Total input size = %s, output size = %s" % (inStr, outStr))
     return
 
 
@@ -442,6 +453,7 @@ def checkURL(url, codemax=200, timeout=3.0):
 
     """
     import requests
+    import logging
 
     retval = False
     try:
@@ -490,7 +502,12 @@ def promptYesNo(msg='', default='n'):
     return retval
 
 
-def modifyFilename(fname, prepend='', append=''):
+def modifyFilename(*args, **kwargs):
+    utils.dep_warn("modifyFilename", newname="modify_filename")
+    return modify_filename(*args, **kwargs)
+
+
+def modify_filename(fname, prepend='', append=''):
     """Modify the given filename.
 
     Arguments
@@ -713,7 +730,7 @@ def modify_exists(fname, max=1000):
         path = './'
     # construct regex for modified files
     #     look specifically for `prec`-digit numbers at the end of the filename
-    regex = modifyFilename(filename, append='_([0-9]){{{:d}}}'.format(prec))
+    regex = modify_filename(re.escape(filename), append='_([0-9]){{{:d}}}'.format(prec))
     matches = sorted([ff for ff in os.listdir(path) if re.search(regex, ff)])
     # If there are matches, find the highest file-number in the matches
     if len(matches):
@@ -735,10 +752,12 @@ def modify_exists(fname, max=1000):
 
     # Construct new filename
     # ----------------------
-    newName = modifyFilename(fname, append='_{0:0{1:d}d}'.format(num, prec))
-    # New filename shouldnt exist; if it does, raise error
+    newName = modify_filename(fname, append='_{0:0{1:d}d}'.format(num, prec))
+    # New filename shouldnt exist; if it does, raise warning
     if os.path.exists(newName):
-        raise RuntimeError("New filename '{}' already exists.".format(newName))
+        # raise RuntimeError("New filename '{}' already exists.".format(newName))
+        warnings.warn("New filename '{}' already exists.".format(newName))
+        return modify_exists(newName)
 
     return newName
 
@@ -749,7 +768,7 @@ def iterable_notstring(var):
     return not isinstance(var, six.string_types) and isinstance(var, collections.Iterable)
 
 
-def str_format_dict(jdict):
+def str_format_dict(jdict, **kwargs):
     """Pretty-format a dictionary into a nice looking string using the `json` package.
 
     Arguments
@@ -763,8 +782,10 @@ def str_format_dict(jdict):
         Nicely formatted string.
 
     """
+    kwargs.setdefault('sort_keys', True)
+    kwargs.setdefault('indent', 4)
     import json
-    jstr = json.dumps({'4': 5, '6': 7}, sort_keys=True, indent=4, separators=(',', ': '))
+    jstr = json.dumps(jdict, separators=(',', ': '), **kwargs)
     return jstr