+++ zcode/inout/log.py
@@ -6,42 +6,54 @@ Classes
 
 Functions
 ---------
--   getLogger                - Create a logger object which logs to file and or stdout stream.
--   defaultLogger            - Create a ``logging.Logger`` object which logs to the out stream.
+-   get_logger                - Create a logger object which logs to file and or stdout stream.
+-   default_logger            - Create a ``logging.Logger`` object which logs to the out stream.
 
 """
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from datetime import datetime
 import logging
+import shutil
 import inspect
 import numpy as np
 
-__all__ = ['IndentFormatter', 'getLogger', 'defaultLogger']
+from . import inout_core
+from .. import utils
+
+__all__ = ['IndentFormatter', 'get_logger', 'default_logger',
+           # DEPRECATED:
+           'getLogger', 'defaultLogger']
 
 
 class IndentFormatter(logging.Formatter):
     """Logging formatter where the depth of the stack sets the message indentation level.
     """
 
-    def __init__(self, fmt=None, datefmt=None):
-        logging.Formatter.__init__(self, fmt, datefmt)
+    def __init__(self, fmt=None, format_date=None):
+        logging.Formatter.__init__(self, fmt, format_date)
         self.baseline = None
 
     def format(self, rec):
         stack = inspect.stack()
-        if self.baseline is None: self.baseline = len(stack)
-        indent = (len(stack)-self.baseline)
+        if (self.baseline is None) or (len(stack) < self.baseline):
+            self.baseline = len(stack)
+        indent = (len(stack) - self.baseline)
         addSpace = ((indent > 0) & (not rec.msg.startswith(" -")))
-        rec.indent = ' -'*indent + ' '*addSpace
+        rec.indent = (' -' * indent) + (' ' * addSpace)
         out = logging.Formatter.format(self, rec)
         del rec.indent
         return out
 
 
-def getLogger(name, strFmt=None, fileFmt=None, dateFmt=None,
-              strLevel=logging.WARNING, fileLevel=logging.DEBUG,
-              tofile=None, tostr=True):
+def getLogger(*args, **kwargs):
+    utils.dep_warn("getLogger", newname="get_logger")
+    return get_logger(*args, **kwargs)
+
+
+def get_logger(name, format_stream=None, format_file=None, format_date=None,
+               level_stream=logging.WARNING, level_file=logging.DEBUG,
+               tofile=None, tostr=True):
     """Create a standard logger object which logs to file and or stdout stream.
 
     If logging to output stream (stdout) is enabled, an `IndentFormatter` object is used.
@@ -50,15 +62,15 @@ def getLogger(name, strFmt=None, fileFmt=None, dateFmt=None,
     ---------
     name : str,
         Handle for this logger, must be distinct for a distinct logger.
-    strFmt : str or `None`,
+    format_stream : str or `None`,
         Format of log messages to stream (stdout).  If `None`, default settings are used.
-    fileFmt : str or `None`,
+    format_file : str or `None`,
         Format of log messages to file.  If `None`, default settings are used.
-    dateFmt : str or `None`
+    format_date : str or `None`
         Format of time stamps to stream and/or file.  If `None`, default settings are used.
-    strLevel : int,
+    level_stream : int,
         Logging level for stream.
-    fileLevel : int,
+    level_file : int,
         Logging level for file.
     tofile : str or `None`,
         Filename to log to (turned off if `None`).
@@ -83,27 +95,27 @@ def getLogger(name, strFmt=None, fileFmt=None, dateFmt=None,
     logger.propagate = 0
 
     # Determine and Set Logging Levels
-    if fileLevel is None:
-        fileLevel = logging.DEBUG
-    if strLevel is None:
-        strLevel = logging.WARNING
+    if level_file is None:
+        level_file = logging.DEBUG
+    if level_stream is None:
+        level_stream = logging.WARNING
     #     Logger object must be at minimum level (`np` int doesnt work, need regular int)
-    logger.setLevel(int(np.min([fileLevel, strLevel]).astype(int)))
+    logger.setLevel(int(np.min([level_file, level_stream]).astype(int)))
 
-    if dateFmt is None:
-        dateFmt = '%Y/%m/%d %H:%M:%S'
+    if format_date is None:
+        format_date = '%Y/%m/%d %H:%M:%S'
 
     # Log to file
     # -----------
     if tofile is not None:
-        if fileFmt is None:
-            fileFmt = "%(asctime)s %(levelname)8.8s [%(filename)20.20s:"
-            fileFmt += "%(funcName)-20.20s]%(indent)s%(message)s"
+        if format_file is None:
+            format_file = "%(asctime)s %(levelname)8.8s [%(filename)20.20s:"
+            format_file += "%(funcName)-20.20s]%(indent)s%(message)s"
 
-        fileFormatter = IndentFormatter(fileFmt, datefmt=dateFmt)
+        fileFormatter = IndentFormatter(format_file, format_date=format_date)
         fileHandler = logging.FileHandler(tofile, 'w')
         fileHandler.setFormatter(fileFormatter)
-        fileHandler.setLevel(fileLevel)
+        fileHandler.setLevel(level_file)
         logger.addHandler(fileHandler)
         #     Store output filename to `logger` object
         logger.filename = tofile
@@ -111,26 +123,45 @@ def getLogger(name, strFmt=None, fileFmt=None, dateFmt=None,
     # Log To stdout
     # -------------
     if tostr:
-        if strFmt is None:
-            strFmt = "%(indent)s%(message)s"
+        if format_stream is None:
+            format_stream = "%(indent)s%(message)s"
 
-        strFormatter = IndentFormatter(strFmt, datefmt=dateFmt)
+        strFormatter = IndentFormatter(format_stream, format_date=format_date)
         strHandler = logging.StreamHandler()
         strHandler.setFormatter(strFormatter)
-        strHandler.setLevel(strLevel)
+        strHandler.setLevel(level_stream)
         logger.addHandler(strHandler)
 
-    # Add a `raise` method to both log an error and raise one
-    # -------------------------------------------------------
-    def _raise(self, msg, error=RuntimeError):
+    # Add a `raise_error` method to both log an error and raise one
+    # -------------------------------------------------------------
+    def _raise_error(self, msg, error=RuntimeError):
+        """Log an error message and raise an error.
+        """
         self.error(msg)
         raise error(msg)
-    # Not entirely sure why this works, but it seems to
-    logger.raise_error = _raise.__get__(logger)
+    logger.raise_error = _raise_error.__get__(logger)
 
     # Add a `after` method to log how long something took
     # ---------------------------------------------------
-    def _after(self, msg, beg, beg_all=None, lvl=logging.INFO):
+    logger._after_lvl = logging.INFO
+
+    def _after(self, msg, beg, beg_all=None, lvl=None):
+        """Log a message and include a report of duration using `datetime`.
+
+        Arguments
+        ---------
+        msg : str
+            Message to log
+        beg : `datetime.Datetime`
+            Datetime of the start of operation (reported duration is `datetime.now() - beg`
+        beg_all : `datetime.Datetime`
+            Datetime of a different start point, duration is given in a parenthesis
+        lvl : int
+            Logging level, default is given by the `_after_lvl` attribute.
+
+        """
+        if lvl is None:
+            lvl = self._after_lvl
         _str = "{} after {}".format(msg, datetime.now()-beg)
         if beg_all is not None:
             _str += " ({})".format(datetime.now()-beg_all)
@@ -138,10 +169,50 @@ def getLogger(name, strFmt=None, fileFmt=None, dateFmt=None,
     # Not entirely sure why this works, but it seems to
     logger.after = _after.__get__(logger)
 
+    # Add a `copy_file` method to copy logfile to the given destination
+    # -----------------------------------------------------------------
+    def _copy(self, dest, modify_exists=False):
+        """Copy the curent output logfile to a new destination.
+        """
+        if modify_exists:
+            dest = inout_core.modify_exists(dest)
+        inout_core.check_path(dest)
+        shutil.copy(self.filename, dest)
+    # Not entirely sure why this works, but it seems to
+    logger.copy = _copy.__get__(logger)
+
+    # Add a `raise_error` method to both log an error and raise one
+    # -------------------------------------------------------------
+    logger._frac_lvl = logging.INFO
+
+    def _frac(self, num, den, prep=None, post=None, lvl=None):
+        """Log information about a fraction, "[{prep} ]{}/{} = {}[ {post}]".
+        """
+        _if = '5d'
+        _ff = '.4f'
+        if lvl is None:
+            lvl = self._frac_lvl
+        _str = ""
+        if prep is not None:
+            _str += "{} ".format(prep)
+        _str += "{0:{i}}/{1:{i}} = {2:{f}}".format(num, den, 1.0*num/den, i=_if, f=_ff)
+        if post is not None:
+            _str += " {}".format(post)
+        self.log(lvl, _str)
+    logger.frac = _frac.__get__(logger)
+
+    for lvl in ['DEBUG', 'INFO', 'WARNING', 'ERROR']:
+        setattr(logger, lvl, getattr(logging, lvl))
+
     return logger
 
 
-def defaultLogger(logger=None, verbose=False, debug=False):
+def defaultLogger(*args, **kwargs):
+    utils.dep_warn("defaultLogger", newname="default_logger")
+    return default_logger(*args, **kwargs)
+
+
+def default_logger(logger=None, verbose=False, debug=False):
     """Create a basic ``logging.Logger`` object which logs to the out stream.
 
     Arguments
@@ -161,7 +232,8 @@ def defaultLogger(logger=None, verbose=False, debug=False):
 
     """
 
-    if isinstance(logger, logging.Logger): return logger
+    if isinstance(logger, logging.Logger):
+        return logger
 
     import numbers
 
@@ -175,5 +247,5 @@ def defaultLogger(logger=None, verbose=False, debug=False):
         else:
             level = logging.WARNING
 
-    logger = getLogger(None, strLevel=level, tostr=True)
+    logger = get_logger(None, level_stream=level, tostr=True)
     return logger